+++ NEWS.md
@@ -1,8 +1,10 @@
 # did 2.2.0
 
-  * added wrapper function for HonestDiD package
+  * Added wrapper function for HonestDiD package
   
   * Fix bug for setups where `gname` is not contained in `tname` (but is in the `tname` range)
+  
+  * Fix bug for including too many groups with universal base period in pre-treatment periods
 
 # did 2.1.1